voice_chatbot/requirements.txt
SpeechRecognition==3.10.0
google-cloud-dialogflow==2.23.1
google-generativeai==0.3.1
gTTS==2.3.2
playsound==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
sympy==1.12
matplotlib==3.7.1
flask==3.0.0
elevenlabs==0.2.27